% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{data}
\alias{data}
\title{data}
\usage{
data(data)
}
\arguments{
\item{data}{the name of the built-in dataset}
}
\value{
the built-in dataset
}
\description{
load a built-in dataset
}
\details{
DETAILS
#' @examples 
\dontrun{
data(NYCMafiaNodes) 
 }
}
