% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{edgeScanMatrix}
\alias{edgeScanMatrix}
\title{edgeScanMatrix}
\usage{
edgeScanMatrix(
  nodes,
  edges,
  thres,
  matrix,
  min = 3,
  weighted = FALSE,
  bipartite = FALSE
)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{thres}{threshold (e.g., distance, travel time) to calculate the number of edges, given the user-defined matrix}

\item{matrix}{a user-defined full matrix, including all pairs of nodes. The matrix's column and row names are consistent with nodes' labels. The cell values can be distance, travel time and so on.}

\item{min}{(optional) minimum number of nodes in the searching window}

\item{weighted}{(optional) boolean value of whether a weighted column has been included.}

\item{bipartite}{(optional) boolean value of whether the data is a bipartite network}
}
\value{
a list of two dataframes. The first R datafrmae contains a column of node label, and a column of heat associated with the node. The second R dataframe contains the edge pairs and a boolean column indicating whether the edge is within the scanning window.
}
\description{
calculate number of edges per node in a given threshold from a user-defined matrix.
}
\details{
DETAILS
}
