% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{edgeScanRadius}
\alias{edgeScanRadius}
\title{EdgeScanRadius}
\usage{
edgeScanRadius(nodes, edges, maxRadius, min = 3)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{maxRadius}{radius in the unit of coordinates of search window}

\item{min}{minimum number of nodes in the distance window}
}
\value{
a R dataframe that contains a column of node label, and a column of heat associated with the node
}
\description{
edgeScan - generates heatmap of edges within maxRadius of every node in a graph
}
\details{
DETAILS
}
