% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{GlobalFlatteningRatio}
\alias{GlobalFlatteningRatio}
\title{Global Network Flattening Ration}
\usage{
GlobalFlatteningRatio(nodes, edges, iter)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{iter}{number of iterations to shuffle the order of nodes for various configuration of G_bar}
}
\value{
a numeric value that is the ratio of sum of distance between G_bar and G
}
\description{
calculate global network flattening ratio for nodes.
}
\details{
DETAILS
}
