% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{processEdge}
\alias{processEdge}
\title{processEdge}
\usage{
processEdge(data, source_name, target_name, weight_name = NA)
}
\arguments{
\item{data}{user_edges, which should be a R dataframe that contains a column of source (string of node label) and a column of target (string of node label)}

\item{source_name}{column name (string) that contains the source node labels}

\item{target_name}{column name (string) that contains the target node labels}

\item{weight_name}{(optional) column name (float) that contains edge weight if available}
}
\value{
a list of lists. Each sublist contains all info of an edge
}
\description{
convert a R dataframe with two columns (Source name and Target name) to correct formats for following algorithms
}
\details{
DETAILS
}
\examples{
\dontrun{
edges = processEdge(user_edges, 'Source', 'Target')
}
}
