% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools (Xiaos-MBP.attlocal.net's
%   conflicted copy 2022-01-28).R, R/oxy-SSNtools.R
\name{processNode}
\alias{processNode}
\title{processNode}
\usage{
processNode(data, label_name, lon_name, lat_name, bipartite_name = NA)

processNode(data, label_name, lon_name, lat_name, bipartite_name = NA)
}
\arguments{
\item{data}{user_nodes, which should be a R dataframe that contains a column of node label, a column of node longitude, and a column of node latitude}

\item{label_name}{column name (string) that contains the node labels}

\item{lon_name}{column name (float) that contains the node longitude}

\item{lat_name}{column name (float) that contains the node latitude}

\item{bipartite_name}{(optional) column name (binary) that contains which set the node is in if available. 1 represents the set that the heat is visualized on.}
}
\value{
a list of named lists, in which the name for each list is the node label. The sublist contains all info of a node

a list of named lists, in which the name for each list is the node label. The sublist contains all info of a node
}
\description{
process user_nodes to correct format for hotspot detection algorithm

process user_nodes to correct format for hotspot detection algorithm
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
nodes = processNode(user_nodes, 'label', 'lon', 'lat')
nodes[[1]] 
nodes[['YOUR NODE LABEL']][['lon']]
}
\dontrun{
nodes = processNode(user_nodes, 'label', 'lon', 'lat')
nodes[[1]] 
nodes[['YOUR NODE LABEL']][['lon']]
}
}
