% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools (Xiaos-MBP.attlocal.net's
%   conflicted copy 2022-01-28).R, R/oxy-SSNtools.R
\name{edgeScanKNearest}
\alias{edgeScanKNearest}
\title{edgeScanKNearest}
\usage{
edgeScanKNearest(nodes, edges, k, weighted = FALSE, bipartite = FALSE)

edgeScanKNearest(nodes, edges, k, weighted = FALSE, bipartite = FALSE)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{k}{number of nodes in search window}

\item{weighted}{(optional) boolean value of whether a weighted column has been included.}

\item{bipartite}{(optional) boolean value of whether the data is a bipartite network}
}
\value{
a R dataframe that contains a column of node label, and a column of heat associated with the node

a list of two dataframes. The first R datafrmae contains a column of node label, and a column of heat associated with the node. The second R dataframe contains the edge pairs and a boolean column indicating whether the edge is within the scanning window.
}
\description{
generates heatmap of edges between k nearest nodes to each independent node in a graph

generates heatmap of edges between k nearest nodes to each independent node in a graph
}
\examples{
\dontrun{
edgeScanKNearest(nodes, edges, 10)
}
\dontrun{
edgeScanKNearest(nodes, edges, 5)
}
}
