% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{edgeScanKNearest}
\alias{edgeScanKNearest}
\title{edgeScanKNearest}
\usage{
edgeScanKNearest(nodes, edges, k)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{k}{number of nodes in search window}
}
\value{
a R dataframe that contains a column of node label, and a column of heat associated with the node
}
\description{
generates heatmap of edges between k nearest nodes to each independent node in a graph
}
\examples{
\dontrun{
edgeScanKNearest(nodes, edges, 10)
}
}
