% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{edgeScanManhattan}
\alias{edgeScanManhattan}
\title{edgeScanManhattan}
\usage{
edgeScanManhattan(
  nodes,
  edges,
  radius,
  min = 3,
  weighted = FALSE,
  bipartite = FALSE
)
}
\arguments{
\item{nodes}{nodes of graph (a list of named lists)}

\item{edges}{edges of graph (a list of lists)}

\item{radius}{radius in the unit of coordinates of search window (Manhattan distance)}

\item{min}{(optional) minimum number of nodes in the searching window}

\item{weighted}{(optional) boolean value of whether a weighted column has been included.}

\item{bipartite}{(optional) boolean value of whether the data is a bipartite network}
}
\value{
a list of two dataframes. The first R datafrmae contains a column of node label, and a column of heat associated with the node. The second R dataframe contains the edge pairs and a boolean column indicating whether the edge is within the scanning window.
}
\description{
calculate number of edges for each independent node in a graph in a range of manhattan distance
}
\details{
DETAILS
}
