% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNHotSpot.R
\name{edgeScanManhattan}
\alias{edgeScanManhattan}
\title{FUNCTION_TITLE}
\usage{
edgeScanManhattan(nodes, edges, distance)
}
\arguments{
\item{nodes}{PARAM_DESCRIPTION}

\item{edges}{PARAM_DESCRIPTION}

\item{distance}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
