% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNHotSpot.R
\name{NDScanKNearest}
\alias{NDScanKNearest}
\title{FUNCTION_TITLE}
\usage{
NDScanKNearest(nodes, edges, k)
}
\arguments{
\item{nodes}{PARAM_DESCRIPTION}

\item{edges}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
