% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{NDScanRadius}
\alias{NDScanRadius}
\title{FUNCTION_TITLE}
\usage{
NDScanRadius(nodes, edges, radius, min = 3)
}
\arguments{
\item{nodes}{PARAM_DESCRIPTION}

\item{edges}{PARAM_DESCRIPTION}

\item{radius}{PARAM_DESCRIPTION}

\item{min}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
