% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-SSNtools.R
\name{processNode}
\alias{processNode}
\title{processNode}
\usage{
processNode(data, label_name, lon_name, lat_name)
}
\arguments{
\item{data}{user_nodes, which should be a R dataframe that contains a column of node label, a column of node longitude, and a column of node latitude}

\item{label_name}{column name (string) that contains the node labels}

\item{lon_name}{column name (float) that contains the node longitude}

\item{lat_name}{column name (float) that contains the node latitude}
}
\value{
a list of named lists, in which the name for each list is the node label. The sublist contains all info of a node
}
\description{
process user_nodes to correct format for hotspot detection algorithm
}
\details{
DETAILS
}
\examples{
\dontrun{
nodes = processNode(user_nodes, 'label', 'lon', 'lat')
nodes[[1]] 
nodes[['YOUR NODE LABEL']][['lon']]
}
}
